Bookshop/main/refactor services (#15)
* Add Order from cart method

* Create AddFromWishList method

* Global refactoring.

* Remove WishListItem part

* Remove transactions

* Refactor Services and thir controllers

* Add pagination for product

* Add Ordering for products

* Fix logger message

* Refactor ProductService

* Refactor some services and controllers.
Add MockBillingService.

* Add mock data in Order

* Add payment creation

* Refactor order service

* Refactor

* Refactor order placement part.

* Add OrderProduct
Comlete entites configurations

* Add migration

* Refactor Order related part

* Add migration

* Refactor Code

* Add page validation

* Add multiple inserting in order placeOrder method
recreate migrations

* Fix orderFromCart method

* Refactor code

* Change InvoiceGetById

* Refactor

* Change CardDetatils property

* Add in order get methods paymant and invoice id

---------

+++ BookShop.Services/Models/CartItemModels/CartItemModel.cs
@@ -4,5 +4,5 @@ public class CartItemModel
 {
     public long Id { get; set; }
     public long ProductId { get; set; }
-    public long Count { get; set; }
+    public int Count { get; set; }
 }
